+++ app/controllers/books_controller.rb
@@ -40,24 +40,33 @@ class BooksController < ApplicationController
   def destroy # DELETE /books/:id
     user = current_user
     book = user.books.find(params[:id])
-    book.destroy
-    flash[:success] = "Book successfully deleted."
+    if user.owns_book?(book)
+      book.destroy
+      flash[:success] = "Book successfully deleted."
+    else
+      # raise the same error as if the book couldnt be found
+      raise ActiveRecord::RecordNotFound
+    end
     redirect_to user, status: :see_other
   end
 
   def new_book_user # POST /bookuser/new
     user = current_user
     book = user.books.find(params[:book_id])
-    begin
-      user_book_was_shared_with = User.find_by(name: params[:name])
-      user_book_was_shared_with.books << book
-      role_id = params[:role_id].to_i
-      book_user = book.book_users.where(["book_id = ? and user_id = ?", book.id, user_book_was_shared_with.id]).first
-      book_user.role_id = role_id if (1...3).cover?(role_id)
-      book_user.save
-      flash[:success] = "Book successfully shared with #{user_book_was_shared_with.name}."
-    rescue
-      flash[:error] = "User was not found; book not shared."
+    if user.owns_book?(book)
+      begin
+        user_book_was_shared_with = User.find_by(name: params[:name])
+        user_book_was_shared_with.books << book
+        role_id = params[:role_id].to_i
+        book_user = book.book_users.where(["book_id = ? and user_id = ?", book.id, user_book_was_shared_with.id]).first
+        book_user.role_id = role_id if (1...3).cover?(role_id)
+        book_user.save
+        flash[:success] = "Book successfully shared with #{user_book_was_shared_with.name}."
+      rescue
+        flash[:error] = "User was not found; book not shared."
+      end
+    else
+      flash[:error] = "You may not share a book that you do not own."
     end
     redirect_to edit_book_path(book)
   end
